refactor: 삭제 기능 리팩토링

+++ src/main/java/qna/domain/Question.java
@@ -1,7 +1,8 @@
 package qna.domain;
 
+import qna.CannotDeleteException;
+
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -71,10 +72,6 @@ public class Question extends AbstractEntity {
         return writer.equals(loginUser);
     }
 
-    public boolean isAllAnswerOwner() {
-        return answers.isOwner(writer);
-    }
-
     public Question setDeleted(boolean deleted) {
         this.deleted = deleted;
         return this;
@@ -88,10 +85,15 @@ public class Question extends AbstractEntity {
         return answers.getAnswers();
     }
 
-    public List<DeleteHistory> delete() {
+    public List<DeleteHistory> delete(User loginUser) throws CannotDeleteException {
+        if (!isOwner(loginUser)) {
+            throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
+        }
+
         this.deleted = true;
         DeleteHistories deleteHistories = new DeleteHistories(this);
-        answers.delete(deleteHistories);
+
+        answers.delete(loginUser, deleteHistories);
         return deleteHistories.getDeleteHistories();
     }
